move to precompile methods
