+++ precompiles/ArbAggregator.go
@@ -22,7 +22,7 @@ type ArbAggregator struct {
 
 // Gets an account's preferred aggregator
 func (con ArbAggregator) GetPreferredAggregator(c ctx, evm mech, address addr) (prefAgg addr, isDefault bool, err error) {
-	l1p := c.state.L1PricingState()
+	l1p := c.State.L1PricingState()
 	maybePrefAgg, err := l1p.UserSpecifiedAggregator(address)
 	if err != nil {
 		return common.Address{}, false, err
@@ -43,22 +43,22 @@ func (con ArbAggregator) SetPreferredAggregator(c ctx, evm mech, prefAgg addr) e
 	if prefAgg != (common.Address{}) {
 		maybePrefAgg = &prefAgg
 	}
-	return c.state.L1PricingState().SetUserSpecifiedAggregator(c.caller, maybePrefAgg)
+	return c.State.L1PricingState().SetUserSpecifiedAggregator(c.caller, maybePrefAgg)
 }
 
 // Gets the chain's default aggregator
 func (con ArbAggregator) GetDefaultAggregator(c ctx, evm mech) (addr, error) {
-	return c.state.L1PricingState().DefaultAggregator()
+	return c.State.L1PricingState().DefaultAggregator()
 }
 
 // Sets the chain's default aggregator (caller must be the current default aggregator, its fee collector, or an owner)
 func (con ArbAggregator) SetDefaultAggregator(c ctx, evm mech, newDefault addr) error {
-	l1State := c.state.L1PricingState()
+	l1State := c.State.L1PricingState()
 	defaultAgg, err := l1State.DefaultAggregator()
 	if err != nil {
 		return err
 	}
-	allowed, err := accountIsAggregatorOrCollectorOrOwner(c.caller, defaultAgg, c.state)
+	allowed, err := accountIsAggregatorOrCollectorOrOwner(c.caller, defaultAgg, c.State)
 	if err != nil {
 		return err
 	}
@@ -70,37 +70,37 @@ func (con ArbAggregator) SetDefaultAggregator(c ctx, evm mech, newDefault addr) 
 
 // Get the aggregator's compression ratio, measured in basis points
 func (con ArbAggregator) GetCompressionRatio(c ctx, evm mech, aggregator addr) (uint64, error) {
-	ratio, err := c.state.L1PricingState().AggregatorCompressionRatio(aggregator)
+	ratio, err := c.State.L1PricingState().AggregatorCompressionRatio(aggregator)
 	return uint64(ratio), err
 }
 
 // Set the aggregator's compression ratio, measured in basis points
 func (con ArbAggregator) SetCompressionRatio(c ctx, evm mech, aggregator addr, newRatio uint64) error {
-	allowed, err := accountIsAggregatorOrCollectorOrOwner(c.caller, aggregator, c.state)
+	allowed, err := accountIsAggregatorOrCollectorOrOwner(c.caller, aggregator, c.State)
 	if err != nil {
 		return err
 	}
 	if !allowed {
 		return errors.New("Only an aggregator (or its fee collector / chain owner) may change its compression ratio")
 	}
-	return c.state.L1PricingState().SetAggregatorCompressionRatio(aggregator, arbmath.Bips(newRatio))
+	return c.State.L1PricingState().SetAggregatorCompressionRatio(aggregator, arbmath.Bips(newRatio))
 }
 
 // Gets an aggregator's fee collector
 func (con ArbAggregator) GetFeeCollector(c ctx, evm mech, aggregator addr) (addr, error) {
-	return c.state.L1PricingState().AggregatorFeeCollector(aggregator)
+	return c.State.L1PricingState().AggregatorFeeCollector(aggregator)
 }
 
 // Sets an aggregator's fee collector (caller must be the aggregator, its fee collector, or an owner)
 func (con ArbAggregator) SetFeeCollector(c ctx, evm mech, aggregator addr, newFeeCollector addr) error {
-	allowed, err := accountIsAggregatorOrCollectorOrOwner(c.caller, aggregator, c.state)
+	allowed, err := accountIsAggregatorOrCollectorOrOwner(c.caller, aggregator, c.State)
 	if err != nil {
 		return err
 	}
 	if !allowed {
 		return errors.New("Only an aggregator (or its fee collector / chain owner) may change its fee collector")
 	}
-	return c.state.L1PricingState().SetAggregatorFeeCollector(aggregator, newFeeCollector)
+	return c.State.L1PricingState().SetAggregatorFeeCollector(aggregator, newFeeCollector)
 }
 
 // Gets an aggregator's current fixed fee to submit a tx
