+++ precompiles/ArbBLS.go
@@ -24,12 +24,12 @@ func (con ArbBLS) GetPublicKey(c ctx, evm mech, address addr) (huge, huge, huge,
 
 // Associate an AltBN128 public key with the caller's address
 func (con ArbBLS) RegisterAltBN128(c ctx, evm mech, x0, x1, y0, y1 huge) error {
-	return c.state.BLSTable().RegisterLegacyPublicKey(c.caller, x0, x1, y0, y1)
+	return c.State.BLSTable().RegisterLegacyPublicKey(c.caller, x0, x1, y0, y1)
 }
 
 // Get the AltBN128 public key associated with an address (revert if there isn't one)
 func (con ArbBLS) GetAltBN128(c ctx, evm mech, address addr) (huge, huge, huge, huge, error) {
-	return c.state.BLSTable().GetLegacyPublicKey(address)
+	return c.State.BLSTable().GetLegacyPublicKey(address)
 }
 
 // Associate a BLS 12-381 public key with the caller's address
@@ -38,12 +38,12 @@ func (con ArbBLS) RegisterBLS12381(c ctx, evm mech, keyBuf []byte) error {
 	if err != nil {
 		return err
 	}
-	return c.state.BLSTable().RegisterBLS12381PublicKey(c.caller, key)
+	return c.State.BLSTable().RegisterBLS12381PublicKey(c.caller, key)
 }
 
 // Get the BLS 12-381 public key associated with an address (revert if there isn't one)
 func (con ArbBLS) GetBLS12381(c ctx, evm mech, address addr) ([]byte, error) {
-	pubKey, err := c.state.BLSTable().GetBLS12381PublicKey(address)
+	pubKey, err := c.State.BLSTable().GetBLS12381PublicKey(address)
 	if err != nil {
 		return nil, err
 	}
