+++ precompiles/ArbOwner.go
@@ -22,7 +22,7 @@ type ArbOwner struct {
 
 // Add account as a chain owner
 func (con ArbOwner) AddChainOwner(c ctx, evm mech, newOwner addr) error {
-	return c.state.ChainOwners().Add(newOwner)
+	return c.State.ChainOwners().Add(newOwner)
 }
 
 // Remove account from the list of chain owners
@@ -31,80 +31,80 @@ func (con ArbOwner) RemoveChainOwner(c ctx, evm mech, addr addr) error {
 	if !member {
 		return errors.New("tried to remove non-owner")
 	}
-	return c.state.ChainOwners().Remove(addr)
+	return c.State.ChainOwners().Remove(addr)
 }
 
 // See if the account is a chain owner
 func (con ArbOwner) IsChainOwner(c ctx, evm mech, addr addr) (bool, error) {
-	return c.state.ChainOwners().IsMember(addr)
+	return c.State.ChainOwners().IsMember(addr)
 }
 
 // Retrieves the list of chain owners
 func (con ArbOwner) GetAllChainOwners(c ctx, evm mech) ([]common.Address, error) {
-	return c.state.ChainOwners().AllMembers()
+	return c.State.ChainOwners().AllMembers()
 }
 
 // Sets the L1 basefee estimate directly, bypassing the autoregression
 func (con ArbOwner) SetL1BaseFeeEstimate(c ctx, evm mech, priceInWei huge) error {
-	return c.state.L1PricingState().SetL1BaseFeeEstimateWei(priceInWei)
+	return c.State.L1PricingState().SetL1BaseFeeEstimateWei(priceInWei)
 }
 
 // Set how slowly ArbOS updates its estimate of the L1 basefee
 func (con ArbOwner) SetL1BaseFeeEstimateInertia(c ctx, evm mech, inertia uint64) error {
-	return c.state.L1PricingState().SetL1BaseFeeEstimateInertia(inertia)
+	return c.State.L1PricingState().SetL1BaseFeeEstimateInertia(inertia)
 }
 
 // Sets the L2 gas price directly, bypassing the pool calculus
 func (con ArbOwner) SetL2BaseFee(c ctx, evm mech, priceInWei huge) error {
-	return c.state.L2PricingState().SetBaseFeeWei(priceInWei)
+	return c.State.L2PricingState().SetBaseFeeWei(priceInWei)
 }
 
 // Sets the minimum base fee needed for a transaction to succeed
 func (con ArbOwner) SetMinimumL2BaseFee(c ctx, evm mech, priceInWei huge) error {
-	return c.state.L2PricingState().SetMinBaseFeeWei(priceInWei)
+	return c.State.L2PricingState().SetMinBaseFeeWei(priceInWei)
 }
 
 // Sets the computational speed limit for the chain
 func (con ArbOwner) SetSpeedLimit(c ctx, evm mech, limit uint64) error {
-	return c.state.L2PricingState().SetSpeedLimitPerSecond(limit)
+	return c.State.L2PricingState().SetSpeedLimitPerSecond(limit)
 }
 
 // Sets the number of seconds worth of the speed limit the gas pool contains
 func (con ArbOwner) SetGasPoolSeconds(c ctx, evm mech, seconds uint64) error {
-	return c.state.L2PricingState().SetGasPoolSeconds(seconds)
+	return c.State.L2PricingState().SetGasPoolSeconds(seconds)
 }
 
 // Set the target fullness in bips the pricing model will try to keep the pool at
 func (con ArbOwner) SetGasPoolTarget(c ctx, evm mech, target uint64) error {
-	return c.state.L2PricingState().SetGasPoolTarget(arbmath.SaturatingCastToBips(target))
+	return c.State.L2PricingState().SetGasPoolTarget(arbmath.SaturatingCastToBips(target))
 }
 
 // Set the extent in bips to which the pricing model favors filling the pool over increasing speeds
 func (con ArbOwner) SetGasPoolWeight(c ctx, evm mech, weight uint64) error {
-	return c.state.L2PricingState().SetGasPoolWeight(arbmath.SaturatingCastToBips(weight))
+	return c.State.L2PricingState().SetGasPoolWeight(arbmath.SaturatingCastToBips(weight))
 }
 
 // Set how slowly ArbOS updates its estimate the amount of gas being burnt per second
 func (con ArbOwner) SetRateEstimateInertia(c ctx, evm mech, inertia uint64) error {
-	return c.state.L2PricingState().SetRateEstimateInertia(inertia)
+	return c.State.L2PricingState().SetRateEstimateInertia(inertia)
 }
 
 // Sets the maximum size a tx (and block) can be
 func (con ArbOwner) SetMaxTxGasLimit(c ctx, evm mech, limit uint64) error {
-	return c.state.L2PricingState().SetMaxPerBlockGasLimit(limit)
+	return c.State.L2PricingState().SetMaxPerBlockGasLimit(limit)
 }
 
 // Gets the network fee collector
 func (con ArbOwner) GetNetworkFeeAccount(c ctx, evm mech) (addr, error) {
-	return c.state.NetworkFeeAccount()
+	return c.State.NetworkFeeAccount()
 }
 
 // Sets the network fee collector
 func (con ArbOwner) SetNetworkFeeAccount(c ctx, evm mech, newNetworkFeeAccount addr) error {
-	return c.state.SetNetworkFeeAccount(newNetworkFeeAccount)
+	return c.State.SetNetworkFeeAccount(newNetworkFeeAccount)
 }
 
 // Upgrades ArbOS to the requested version at the requested timestamp
 func (con ArbOwner) ScheduleArbOSUpgrade(c ctx, evm mech, newVersion uint64, timestamp uint64) error {
-	return c.state.ScheduleArbOSUpgrade(newVersion, timestamp)
+	return c.State.ScheduleArbOSUpgrade(newVersion, timestamp)
 }
